requirements.txt
openai>=1.30.0
chromadb>=0.5.0
openai-whisper>=20231117
yt-dlp>=2024.01.01
sentence-transformers>=3.0
python-dotenv>=1.0.0
rich>=13.0
psutil>=5.9.0
pypdf>=4.0
